requirements.txt
# Core dependencies
torch==2.1.0
torchvision==0.16.0
numpy==1.26.0
pillow==11.0.0

# Diffusion models
diffusers
transformers==4.28.1
accelerate==0.27.2
safetensors==0.4.1
xformers==0.0.22.post7

# 3D processing
open3d==0.19.0
trimesh==4.3.2
pytorch3d==0.7.5
opencv-python==4.8.1.78
scipy==1.11.4
scikit-image==0.22.0

# Deep learning utilities
einops==0.7.0
timm==0.9.12
kornia==0.6.9
flash-attn==2.5.3
bitsandbytes==0.42.0

# Data processing
omegaconf==2.3.0
imageio==2.33.1
pandas==2.1.3
tqdm==4.67.0
pyyaml==6.0.1

# Visualization
matplotlib==3.8.2
plotly==6.3.1

# Training utilities
tensorboard==2.15.1
wandb==0.16.1
pytorch-lightning==1.4.2

# Additional utilities
hydra-core
webdataset==0.2.83
lpips==0.1.4
fire==0.5.0
